Feat: par utils: categorize simple loop IVs start and end values
Currently, SCEV does not provide a characterization of the end
value of a simple loop, just the start and recurrence values.

However, given we have an understanding of the loop structure,
deducing this information is possible, though not clean. It
requires a look at the compare predicate and branch instruction
in much depth.