Refactor: partition heuristic: scan all possible subset merges for best
Previously, we greedily searched the subset DAG from top to bottom, merging
as we went. Now, we scan the subset DAG from top to bottom each time for
the best greedy merge. Runtime has not been significantly impacted due to
the fact that SCC count doesn't grow as quickly as Instruction count