wip+Feat/Bugfix: dgsimplify: stop inlining after invalidating DG.
The work in progress is because we still want to inline, but we
need to run the pass again. The pass now writes to a file whether
it inlined and would like to run again. For the time being, a bash
script determining whether to run opt on the pass again will be
sufficient.