Bugfix: sccdag partition: can merge was not conservative enough in its check
Now though, it is more conservative than it needs to be, discounting
cases where one subset has more than one ancestor, one of which is the
other subset, but none of which share ancestors, therefore forming no
cycle. Eventually, this canMerge helper will need to full-out check for
a cycle to avoid being too conservative.