Refactor: Consolidated DSWPLoopDependenceInfo into LoopDependenceInfo
The structure was unnecessary; all fields DSWPLoopDependenceInfo
contained had nothing to do with information about the loop. They
belonged in the class DSWP where they are now. With that simplification,
all passes interact with the one and only LoopDependenceInfo.

The name of this structure, by the way, is cause for concern. It isn't
solely dependence information that it stores. But that is a discussion
for another day. One task at a time.