OracleAA now parses Free and Alloc dependencies and maps these to Read and Write dependencies, inputting RaW (true) dependencies.
Changed the function results to use at and find instead of operator::[]
so meaning having not entry in the mapping means that the loop
was not instrumented.

Now looks for a .viaconf or .dep file to load in and therefore not
instrument all loops or run the instrumentation if the file is present.