BUGFIX: InductionVariable/SCCDAGAttrs: correct contained instr collection
InductionVariable was falsely claiming that the evolution of the loop header PHI it
was analyzing depended on instructions outside of that loop ; this then
hid outer loop behavior that isn't yet able to be modeled by InductionVariable
from being characterized as such. As a result, SCCDAGAttrs marked SCCs as
ONLY containing induction variable related instructions falsely.